Feat: Se agrega alerta en la validacion de campos obligatorios en el formulario de contactanos y alerta de mantenimiento en el boton de asesor de ventas

+++ javascript/contactanos.js
@@ -1,3 +1,4 @@
+const form = document.querySelector("form.formulario-informacion")
 const nombreUsuario = document.querySelector("input.form-control#nombre");
 const apellidoUsuario = document.querySelector("input.form-control#apellido");
 const emailUsuario = document.querySelector("input.form-control#email");
@@ -34,6 +35,45 @@ function guardarDatosUsuario(){
         opinion: opinionUsuario.value
     }
 
-    localStorage.setItem("dataUsuario", JSON.stringify(dataUsuario))
+    //validacion formulario lleno
+    let elementos_form = form.elements;
+
+    // Itero sobre los elementos del formulario para validar que NO estén vacios
+    for (let i = 0; i < elementos_form.length; i++) {
+
+        if (elementos_form[i].type !== "button" && elementos_form[i].value.trim() === ""){
+            Swal.fire({
+                icon: "warning",
+                title: "Ops...",
+                text: "El formulario no puede estar vacío, por favor llena la información.",
+                customClass: {
+                    icon: 'swal-icon-class',
+                    popup: 'swal-container-class',
+                    title: 'swal-title-class',
+                    htmlContainer: 'swal-html-container-class',
+                    confirmButton: 'swal-confirm-button-class'
+                },
+            });
+            return false;
+        } else {
+            Swal.fire({
+                icon: "success",
+                title: "Hemos recibido tus comentarios",
+                text: "Gracias por preferirnos, tu solicitud se respondera en breve.",
+                customClass: {
+                    popup: 'swal-container-class',
+                    title: 'swal-title-class',
+                    htmlContainer: 'swal-html-container-class',
+                    confirmButton: 'swal-confirm-button-class'
+                },
+            });
+
+            // setea el formulario
+            form.reset();
+
+            localStorage.setItem("dataUsuario", JSON.stringify(dataUsuario))
+            return true
+        }
+    }
 }
 botonEnviar.addEventListener('click', ()=> guardarDatosUsuario())
